feat: added feedback section on home page
